# Performance improvements in `JsonValue`.