--- a/src/libraries/System.Text.Json/src/System/Text/Json/Document/JsonDocument.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Document/JsonDocument.cs
@@ -118,10 +118,14 @@ namespace System.Text.Json
             return _parsedData.GetJsonTokenType(index);
         }
 
-        internal bool GetHasComplexChildren(int index)
+        internal bool ValueIsEscaped(int index, bool isPropertyName)
         {
             CheckNotDisposed();
-            DbRow row = _parsedData.Get(index);
+
+            int matchIndex = isPropertyName ? index - DbRow.Size : index;
+            DbRow row = _parsedData.Get(matchIndex);
+            Debug.Assert(!isPropertyName || row.TokenType is JsonTokenType.PropertyName);
+
             return row.HasComplexChildren;
         }
 
@@ -368,6 +372,16 @@ namespace System.Text.Json
         {
             // The property name is one row before the property value
             return GetString(index - DbRow.Size, JsonTokenType.PropertyName)!;
+        }
+
+        internal ReadOnlySpan<byte> GetPropertyNameRaw(int index)
+        {
+            CheckNotDisposed();
+
+            DbRow row = _parsedData.Get(index - DbRow.Size);
+            Debug.Assert(row.TokenType is JsonTokenType.PropertyName);
+
+            return _utf8Json.Span.Slice(row.Location, row.SizeOrLength);
         }
 
         internal bool TryGetValue(int index, [NotNullWhen(true)] out byte[]? value)