--- a/src/libraries/System.Text.Json/src/System/Text/Json/Document/JsonElement.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Document/JsonElement.cs
@@ -1164,6 +1164,13 @@ namespace System.Text.Json
             return _parent.GetNameOfPropertyValue(_idx);
         }
 
+        internal ReadOnlySpan<byte> GetPropertyNameRaw()
+        {
+            CheckValidInstance();
+
+            return _parent.GetPropertyNameRaw(_idx);
+        }
+
         /// <summary>
         ///   Gets the original input data backing this value, returning it as a <see cref="string"/>.
         /// </summary>
@@ -1199,9 +1206,9 @@ namespace System.Text.Json
             // TODO make public https://github.com/dotnet/runtime/issues/77666
             get
             {
-                Debug.Assert(ValueKind is JsonValueKind.String);
                 CheckValidInstance();
-                return _parent.GetHasComplexChildren(_idx);
+
+                return _parent.ValueIsEscaped(_idx, isPropertyName: false);
             }
         }
 
@@ -1210,9 +1217,135 @@ namespace System.Text.Json
             // TODO make public https://github.com/dotnet/runtime/issues/77666
             get
             {
-                Debug.Assert(ValueKind is JsonValueKind.String);
                 CheckValidInstance();
+
                 return _parent.GetRawValue(_idx, includeQuotes: false).Span;
+            }
+        }
+
+        // TODO make public https://github.com/dotnet/runtime/issues/33388
+        internal static bool DeepEquals(JsonElement left, JsonElement right)
+        {
+            Debug.Assert(left._parent != null);
+            Debug.Assert(right._parent != null);
+
+            JsonValueKind kind = left.ValueKind;
+            if (kind != right.ValueKind)
+            {
+                return false;
+            }
+
+            switch (kind)
+            {
+                case JsonValueKind.Null or JsonValueKind.False or JsonValueKind.True:
+                    return true;
+
+                case JsonValueKind.Number:
+                    // JSON numbers are equal if their raw representations are equal.
+                    return left.GetRawValue().Span.SequenceEqual(right.GetRawValue().Span);
+
+                case JsonValueKind.String:
+                    if (right.ValueIsEscaped)
+                    {
+                        if (left.ValueIsEscaped)
+                        {
+                            // Both values are escaped, force an allocation to unescape the RHS.
+                            return left.ValueEquals(right.GetString());
+                        }
+
+                        // Swap values so that unescaping is handled by the LHS.
+                        (left, right) = (right, left);
+                    }
+
+                    return left.ValueEquals(right.ValueSpan);
+
+                case JsonValueKind.Array:
+                    ArrayEnumerator rightArrayEnumerator = right.EnumerateArray();
+                    foreach (JsonElement leftElement in left.EnumerateArray())
+                    {
+                        if (!rightArrayEnumerator.MoveNext() || !DeepEquals(leftElement, rightArrayEnumerator.Current))
+                        {
+                            return false;
+                        }
+                    }
+
+                    return !rightArrayEnumerator.MoveNext();
+
+                default:
+                    Debug.Assert(kind is JsonValueKind.Object);
+                    ObjectEnumerator leftObjectEnumerator = left.EnumerateObject();
+                    ObjectEnumerator rightObjectEnumerator = right.EnumerateObject();
+
+                    // Two JSON objects are considered equal if they define the same set of properties.
+                    // Start optimistically with sequential comparison, but fall back to unordered
+                    // comparison as soon as a mismatch is encountered.
+
+                    while (leftObjectEnumerator.MoveNext())
+                    {
+                        if (!rightObjectEnumerator.MoveNext())
+                        {
+                            return false;
+                        }
+
+                        JsonProperty leftProp = leftObjectEnumerator.Current;
+                        JsonProperty rightProp = rightObjectEnumerator.Current;
+
+                        if (!NameEquals(leftProp, rightProp))
+                        {
+                            // We have our first mismatch, fall back to unordered comparison.
+                            return UnorderedObjectDeepEquals(leftObjectEnumerator, rightObjectEnumerator);
+                        }
+
+                        if (!DeepEquals(leftProp.Value, rightProp.Value))
+                        {
+                            return false;
+                        }
+                    }
+
+                    return !rightObjectEnumerator.MoveNext();
+
+                    static bool UnorderedObjectDeepEquals(ObjectEnumerator left, ObjectEnumerator right)
+                    {
+                        Dictionary<string, JsonElement> rightElements = new(StringComparer.Ordinal);
+                        do
+                        {
+                            JsonProperty prop = right.Current;
+                            rightElements.TryAdd(prop.Name, prop.Value);
+                        }
+                        while (right.MoveNext());
+
+                        int leftCount = 0;
+                        do
+                        {
+                            JsonProperty prop = left.Current;
+                            if (!rightElements.TryGetValue(prop.Name, out JsonElement rightElement) || !DeepEquals(prop.Value, rightElement))
+                            {
+                                return false;
+                            }
+
+                            leftCount++;
+                        }
+                        while (left.MoveNext());
+
+                        return leftCount == rightElements.Count;
+                    }
+
+                    static bool NameEquals(JsonProperty left, JsonProperty right)
+                    {
+                        if (right.NameIsEscaped)
+                        {
+                            if (left.NameIsEscaped)
+                            {
+                                // Both values are escaped, force an allocation to unescape the RHS.
+                                return left.NameEquals(right.Name);
+                            }
+
+                            // Swap values so that unescaping is handled by the LHS
+                            (left, right) = (right, left);
+                        }
+
+                        return left.NameEquals(right.NameSpan);
+                    }
             }
         }
 
@@ -1312,6 +1445,13 @@ namespace System.Text.Json
             CheckValidInstance();
 
             return _parent.TextEquals(_idx, text, isPropertyName);
+        }
+
+        internal bool ValueIsEscapedHelper(bool isPropertyName)
+        {
+            CheckValidInstance();
+
+            return _parent.ValueIsEscaped(_idx, isPropertyName);
         }
 
         /// <summary>