One of the oddities of `JsonNode` design is that it permits encapsulation of arbitrary .NET objects via the `JsonValue` type. This is known to create performance problems since it's not possible to introspect `JsonValue` instances without serializing and deserializing the underlying object first. The problem is fairly pervasive since JSON primitives such as boolean, numbers and strings are all represented using `JsonValue`.

This PR refactors the implementation of `JsonValue` so that common primitive types and deserialized values backed by `JsonElement` are fully segregated from the implementation that admits arbitrary objects. Consequently this change records [substantial performance improvements](https://gist.github.com/eiriktsarpalis/38cf615c15b2c414b1611fa6ae58cff1):

| Method                   | Branch | Mean         | Ratio | Allocated | Alloc Ratio |
|------------------------- |------- |-------------:|------:|----------:|------------:|
| DeepEquals               | main   | 2,859.165 ns |  1.00 |    4224 B |        1.00 |
| DeepEquals               | PR     |   270.130 ns |  0.09 |         - |        0.00 |
|                          |        |              |       |           |             |
| DeepEquals_JsonElement   | main   |   660.551 ns |  1.00 |     328 B |        1.00 |
| DeepEquals_JsonElement   | PR     |   432.391 ns |  0.65 |         - |        0.00 |
|                          |        |              |       |           |             |
| GetValueKind             | main   |   612.784 ns | 1.000 |     616 B |        1.00 |
| GetValueKind             | PR     |     5.243 ns | 0.009 |         - |        0.00 |
|                          |        |              |       |           |             |
| GetValueKind_JsonElement | main   |    15.557 ns |  1.00 |         - |          NA |
| GetValueKind_JsonElement | PR     |    11.111 ns |  0.71 |         - |          NA |