@eiriktsarpalis 2024-06-19
> Just confirming that we're not losing the ability to store arbitrary objects in `JsonValue`. I actually utilize this in [my JSON-e implementation](https://github.com/json-everything/json-everything/blob/master/src%2FJsonE%2FFunctionDefinition.cs).

We couldn't do that. This change simply isolates that use case from the hot path scenaria.